==> 01 Tweet Scrapper.ipynb <==
# Import libraries
import pandas as pd
from tqdm.notebook import tqdm
import snscrape.modules.twitter as sntwitter

# Scrape tweets related to Bitcoin or BTC
scraper = sntwitter.TwitterSearchScraper("bitcoin or btc since:2021-01-01 until:2023-01-01")

tweets = []

for i, tweet in tqdm(enumerate(scraper.get_items())):
    data = [tweet.date, 
            tweet.id,
            tweet.rawContent, 
            tweet.user.username,
            tweet.user.followersCount, 
            tweet.user.verified, 
            tweet.retweetCount, 
            tweet.likeCount, 
            tweet.quoteCount
           ]
    tweets.append(data)

# Format scraped tweets into DataFrame
tweet_df = pd.DataFrame(tweets, 
                        columns=['date', 'id', 'content', 'username', 
                                 'followers count', 'verified user', 
                                 'retweet count', 'like count', 'quote count'])
tweet_df

# Export scraped tweets into .csv file
tweet_df.to_csv("data/Bitcoin Tweets.csv")
