==> 02 CoinDesk News Scraper.ipynb <==
# Import libraries
import requests
import pandas as pd
from bs4 import BeautifulSoup as bs
from selenium import webdriver
from tqdm import tqdm
from datetime import datetime

### Scraping Sites Methods

# Scrape Markets site
def scrapeMarketsSite(weblink):
    
    req2 = requests.get(weblink)
    soup2 = bs(req2.text, 'html.parser')

    title = soup2.find('div', class_='at-headline').find('h1', class_='typography__StyledTypography-owin6q-0 kVbxLR').text.strip()
    
    date = soup2.find('div', class_='at-created label-with-icon').find('span', class_='typography__StyledTypography-owin6q-0 fUOSEs').text.strip()
    date = datetime.strptime(date[:12], '%b %d, %Y').strftime('%Y-%m-%d')
    
    content =soup2.find('div', class_='contentstyle__StyledWrapper-g5cdrh-0 jsJHBz').findAll('p')
    content = " ".join(list(map(lambda x: x.text.strip(), content)))
    
    return title, date, content

# Scrape Podcasts site
def scrapePodcastSite(weblink):

    req2 = requests.get(weblink)
    soup2 = bs(req2.text, 'html.parser')

    title = soup2.find('div', class_='episode-title').find('h1', class_='typography__StyledTypography-owin6q-0 hjHKEC').text.strip()

    date = soup2.find('div', class_='episode-published-date').find('span', class_='typography__StyledTypography-owin6q-0 bZvpZH').text.strip()
    date = datetime.strptime(date, '%B %d, %Y').strftime('%Y-%m-%d')

    content =soup2.find('div', class_='contentstyle__StyledWrapper-g5cdrh-0 jsJHBz').findAll('p')
    content = " ".join(list(map(lambda x: x.text.strip(), content)))
    
    return title, date, content

# Scrape Video site
def scrapeVideoSite(weblink):

    req2 = requests.get(weblink)
    soup2 = bs(req2.text, 'html.parser')

    title = soup2.find('div', class_='Box-sc-1hpkeeg-0 dNNZxF').find('h1', class_='typography__StyledTypography-owin6q-0 kzGyYF video-title').text.strip()

    date = soup2.find('span', class_='typography__StyledTypography-owin6q-0 fUOSEs video-date').text.strip()
    date = datetime.strptime(date, '%b %d, %Y').strftime('%Y-%m-%d')

    content =soup2.find('div', class_='show-morestyles__ContentWrapper-jwdjew-0 epVFdw').findAll('p')
    content = " ".join(list(map(lambda x: x.text.strip(), content)))
    
    return title, date, content

# Scrape TV site
def scrapeTVSite(weblink):
    
    req2 = requests.get(weblink)
    soup2 = bs(req2.text, 'html.parser')

    title = soup2.find('div', class_='Box-sc-1hpkeeg-0 dNNZxF').find('h1', class_='typography__StyledTypography-owin6q-0 kzGyYF video-title').text.strip()

    date = soup2.find('span', class_='typography__StyledTypography-owin6q-0 fUOSEs video-date').text.strip()
    date = datetime.strptime(date, '%b %d, %Y').strftime('%Y-%m-%d')

    content =soup2.find('div', class_='show-morestyles__ContentWrapper-jwdjew-0 cLoGZO').findAll('p')
    content = " ".join(list(map(lambda x: x.text.strip(), content)))
    
    return title, date, content

# Scrape Consensus site
def scrapeConsensusSite(weblink): 
    
    req2 = requests.get(weblink)
    soup2 = bs(req2.text, 'html.parser')

    title = soup2.find('div', class_='at-headline').find('h1', class_='typography__StyledTypography-owin6q-0 kVbxLR').text.strip()

    date = soup2.find('span', class_='typography__StyledTypography-owin6q-0 fUOSEs').text.strip()
    date = datetime.strptime(date[:12], '%b %d, %Y').strftime('%Y-%m-%d')  

    content =soup2.find('div', class_='contentstyle__StyledWrapper-g5cdrh-0 jsJHBz').findAll('p')
    content = " ".join(list(map(lambda x: x.text.strip(), content)))
    
    return title, date, content

# Scrape Story site
def scrapeStorySite(weblink): 
    
    req2 = requests.get(weblink)
    soup2 = bs(req2.text, 'html.parser')

    title = soup2.find('div', class_='at-headline').find('h1', class_='typography__StyledTypography-owin6q-0 kVbxLR').text.strip()

    date = soup2.find('span', class_='typography__StyledTypography-owin6q-0 fUOSEs').text.strip()
    date = datetime.strptime(date[:12], '%b %d, %Y').strftime('%Y-%m-%d')  

    content =soup2.find('div', class_='at-content-wrapper').findAll('p')
    content = " ".join(list(map(lambda x: x.text.strip(), content)))
    
    return title, date, content

# Scrape Learn site
def scrapeLearnSite(weblink): 
    
    req2 = requests.get(weblink)
    soup2 = bs(req2.text, 'html.parser')

    title = soup2.find('div', class_='at-headline headline2').find('h1', class_='typography__StyledTypography-owin6q-0 fPbJUO').text.strip()

    date = soup2.findAll('meta', property='article:published_time')[0]["content"][:10]

    content =soup2.find('div', class_='at-content-wrapper').findAll('p')
    content = " ".join(list(map(lambda x: x.text.strip(), content)))
    
    return title, date, content

# Scrape Vidoe site 2
def scrapeVideo2Site(weblink):
    
    req2 = requests.get(weblinkk)
    soup2 = bs(req2.text, 'html.parser')

    title = soup2.find('div', class_='at-headline').find('h1', class_='typography__StyledTypography-owin6q-0 kVbxLR').text.strip()

    date = soup2.findAll('meta', property='article:published_time')[0]["content"][:10]

    content =soup2.find('div', class_='at-content-wrapper').findAll('p')
    content = " ".join(list(map(lambda x: x.text.strip(), content)))
    
    return title, date, content

# Scrape TV site 2
def scrapeTV2Site(weblink):
    
    req2 = requests.get(weblink)
    soup2 = bs(req2.text, 'html.parser')

    title = soup2.find('div', class_='Box-sc-1hpkeeg-0 dNNZxF').find('h1', class_='typography__StyledTypography-owin6q-0 kzGyYF video-title').text.strip()

    date = soup2.find('span', class_='typography__StyledTypography-owin6q-0 fUOSEs video-date').text.strip()
    date = datetime.strptime(date, '%b %d, %Y').strftime('%Y-%m-%d')

    content =soup2.find('div', class_='show-morestyles__ContentWrapper-jwdjew-0 fQpATL').findAll('p')
    content = " ".join(list(map(lambda x: x.text.strip(), content)))
    
    return title, date, content

# Scrape Other site
def scrapeOtherSite(weblink): 
    
    req2 = requests.get(weblink)
    soup2 = bs(req2.text, 'html.parser')

    title = soup2.find('div', class_='at-headline').find('h1', class_='typography__StyledTypography-owin6q-0 kVbxLR').text.strip()

    date = soup2.find('span', class_='typography__StyledTypography-owin6q-0 fUOSEs').text.strip()
    try:
        date = datetime.strptime(date[:11], '%b %d, %Y').strftime('%Y-%m-%d')
    except:
        try: 
            date = datetime.strptime(date[:11], '%B %d, %Y').strftime('%Y-%m-%d')
        except:
            date = date

    content =soup2.find('div', class_='at-content-wrapper').findAll('p')
    content = " ".join(list(map(lambda x: x.text.strip(), content)))
    
    return title, date, content

### Scraping

# Create empty dataframe
df = pd.DataFrame({'Title':[], 'Date':[], 'Content':[], 'Page':[]})

# Loop through bitcoin searches and scrape all relevant links
browser=webdriver.Chrome()

for pageNum in tqdm(range(1355)):
    url = f"https://www.coindesk.com/search?s=bitcoin&sort=1&i={pageNum}"
    browser.get(url) #navigate to the page
    innerHTML = browser.execute_script("return document.body.innerHTML")
    
    soup = bs(innerHTML, 'html.parser')
    articles = soup.findAll("div", class_="searchstyles__SearchCardWrapper-ci5zlg-21 hZygnS")
    for article in articles:
        links = article.findAll("a", class_="Box-sc-1hpkeeg-0 hBnhmi")
        if len(links)==2:
            link = links[1]["href"]
            try:
                if link[1:6]=="video":
                    try:
                        title, date, content = scrapeVideoSite("https://www.coindesk.com" + link)
                    except:
                        title, date, content = scrapeVideo2Site("https://www.coindesk.com" + link)
                elif link[1:3]=="tv":
                    try:
                        title, date, content = scrapeTVSite("https://www.coindesk.com" + link)
                    except:
                        title, date, content = scrapeTV2Site("https://www.coindesk.com" + link)
                elif link[1:9]=="podcasts":
                    title, date, content = scrapePodcastSite("https://www.coindesk.com" + link)
                elif link[1:19]=="consensus-magazine":
                    title, date, content = scrapeConsensusSite("https://www.coindesk.com" + link)
                elif link[1:6]=="learn":
                    title, date, content = scrapeLearnSite("https://www.coindesk.com" + link)
                else:
                    title, date, content = scrapeArticle("https://www.coindesk.com" + link)
                df = pd.concat([df, pd.Series({'Title':title, 'Date':date, 'Content':content, 'Page':pageNum}).to_frame().T], ignore_index=True)
            except:
                try:
                    try:
                        title, date, content = scrapeStorySite("https://www.coindesk.com" + link)
                    except:
                        title, date, content = scrapeOtherSite("https://www.coindesk.com" + link)
                    df = pd.concat([df, pd.Series({'Title':title, 'Date':date, 'Content':content, 'Page':pageNum}).to_frame().T], ignore_index=True)
                except:
                    df = pd.concat([df, pd.Series({'Title':link, 'Date':None, 'Content':None, 'Page':pageNum}).to_frame().T], ignore_index=True)
    
    # Save every 100 pages
    # if pageNum % 100 == 0:
    #     df.to_csv("coin desk news " + str(pageNum-100) + " to " + str(pageNum) + ".csv")
    #     df = df.iloc[0:0]

df.to_csv("CoinDesk News.csv")
