==> 03 Tweet Sentiment Analysis.ipynb <==
# Import libararies
import pandas as pd
import numpy as np
from textblob import TextBlob
from wordcloud import WordCloud, STOPWORDS
from PIL import Image
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from langdetect import detect
from nltk.stem import SnowballStemmer
from sklearn.feature_extraction.text import CountVectorizer
import nltk
from transformers import BertTokenizer, BertForSequenceClassification
nltk.download('vader_lexicon')

# Import tweets data
btc_tweets_df = pd.read_csv("data/Bitcoin Tweets.csv")
btc_tweets_df = btc_tweets_df[::-1].reset_index().drop("index", axis=1)
btc_tweets_df['date'] = btc_tweets_df['date'].astype(str).str[:10]
btc_tweets_df['date'] = pd.to_datetime(btc_tweets_df["date"])
btc_tweets_df = btc_tweets_df.sort_values("date")
btc_tweets_df = btc_tweets_df.reset_index().drop("index", axis=1)
btc_tweets_df.head()

# VADER Sentiment

# Conduct VADER sentiment analysis on tweets
neg_scores = []
neu_scores = []
pos_scores = []
comp_scores = []

for t in btc_tweets_df['content']:
    score = SentimentIntensityAnalyzer().polarity_scores(t)
    neg_scores.append(score['neg'])
    neu_scores.append(score['neu'])
    pos_scores.append(score['pos'])
    comp_scores.append(score['compound'])

btc_tweets_df['VADER Neg'] = neg_scores
btc_tweets_df['VADER Neu'] = neu_scores
btc_tweets_df['VADER Pos'] = pos_scores
btc_tweets_df['VADER Comp'] = comp_scores

btc_tweets_df.head()

# Score sentiments based on follower count, like count and retweet count
sentiment_scores = []
for i in range(len(btc_tweets_df)):
    dat = btc_tweets_df.iloc[i]
    s = dat["VADER Comp"] * ((int(dat["followers count"])+1) + (int(dat["like count"])+1) * (int(dat["retweet count"])+1))
    sentiment_scores.append(s)
btc_tweets_df['VADER Score'] = sentiment_scores
btc_tweets_df.head()

# FinBert Sentiment

# Initialize FinBert
finbert = BertForSequenceClassification.from_pretrained('yiyanghkust/finbert-tone',num_labels=3)
tokenizer = BertTokenizer.from_pretrained('yiyanghkust/finbert-tone')

labels = {0:'neutral', 1:'positive',2:'negative'}

# Conduct FinBert sentiment analysis on tweets
sent_val = list()
for tweet in btc_tweets_df['content'].to_list():
    inputs = tokenizer(tweet, return_tensors="pt", padding=True)
    outputs = finbert(**inputs)[0]
   
    val = labels[np.argmax(outputs.detach().numpy())]   
    sent_val.append(val)
    
btc_tweets_df['FinBert Sentiment'] = sent_val

# One-hot encoding of FinBert sentiment labels on tweets
btc_tweets_df['FinBert Pos'] = list(map(lambda x: 1 if x else 0, btc_tweets_df["FinBert Sentiment"]=="positive"))
btc_tweets_df['FinBert Neu'] = list(map(lambda x: 1 if x else 0, btc_tweets_df["FinBert Sentiment"]=="neutral"))
btc_tweets_df['FinBert Neg'] = list(map(lambda x: 1 if x else 0, btc_tweets_df["FinBert Sentiment"]=="negative"))

btc_tweets_df.head()

# Aggregate tweets daily
daily_tweet_count = btc_tweets_df.groupby('date')['content'].count()
daily_like_count = btc_tweets_df.groupby('date')['like count'].sum()
daily_retweet_count = btc_tweets_df.groupby('date')['retweet count'].sum()

avg_daily_vader_neg = btc_tweets_df.groupby('date')['VADER Neg'].mean()
avg_daily_vader_neu = btc_tweets_df.groupby('date')['VADER Neu'].mean()
avg_daily_vader_pos = btc_tweets_df.groupby('date')['VADER Pos'].mean()
avg_daily_vader_comp = btc_tweets_df.groupby('date')['VADER Comp'].mean()
avg_daily_vader_score = btc_tweets_df.groupby('date')['VADER Score'].mean()

daily_finbert_neg_count = btc_tweets_df.groupby('date')['FinBert Neg'].sum()
daily_finbert_neu_count = btc_tweets_df.groupby('date')['FinBert Neu'].sum()
daily_finbert_pos_count = btc_tweets_df.groupby('date')['FinBert Pos'].sum()

daily_total_sent_count = daily_finbert_neg_count + daily_finbert_neu_count + daily_finbert_pos_count
daily_prop_finbert_neg = daily_finbert_neg_count / daily_total_sent_count
daily_prop_finbert_neu = daily_finbert_neu_count / daily_total_sent_count
daily_prop_finbert_pos = daily_finbert_pos_count / daily_total_sent_count

btc_tweets_summary_df = pd.DataFrame({"Tweet count":daily_tweet_count, 
                                      "Retweet count":daily_retweet_count,
                                      "Like count":daily_like_count,
                                      "Avg VADER Neg":avg_daily_vader_neg,
                                      "Avg VADER Neu":avg_daily_vader_neu,
                                      "Avg VADER Pos":avg_daily_vader_pos,
                                      "Avg VADER Comp":avg_daily_vader_comp,
                                      "Avg VADER Score":avg_daily_vader_score,
                                      "FinBert Neg count": daily_finbert_neg_count,
                                      "FinBert Neu count": daily_finbert_neu_count,
                                      "FinBert Pos count": daily_finbert_pos_count,
                                      "FinBert proportion Neg": daily_prop_finbert_neg,
                                      "FinBert proportion Neu": daily_prop_finbert_neu,
                                      "FinBert proportion Pos": daily_prop_finbert_pos,
                                     })

btc_tweets_summary_df.head()

# Export Tweets sentiment
btc_tweets_summary_df.to_csv("Tweet Sentiments.csv")
