==> 04 CoinDesk News Sentiment Analysis.ipynb <==
import pandas as pd
import numpy as np
from textblob import TextBlob
from wordcloud import WordCloud, STOPWORDS
from PIL import Image
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from langdetect import detect
from nltk.stem import SnowballStemmer
from sklearn.feature_extraction.text import CountVectorizer
import spacy
import nltk
from nltk.corpus import stopwords
from transformers import BertTokenizer, BertForSequenceClassification
nltk.download('vader_lexicon')

# Import CoinDesk news data
coindesk_df = pd.read_csv("data/CoinDesk News.csv", encoding='unicode_escape') 
coindesk_df.drop("Unnamed: 0", axis=1, inplace=True)
coindesk_df["Date"] = pd.to_datetime(coindesk_df["Date"])
coindesk_df = coindesk_df.sort_values("Date")
coindesk_df = coindesk_df.reset_index().drop("index", axis=1)
coindesk_df.head()

# FinBert Sentiment

# Text pre-processing
nlp = spacy.load('en_core_web_sm')
stop = stopwords.words('english')

coindesk_df['Cleaned title'] = coindesk_df['Title'].apply(lambda x: " ".join(x.lower() for x in str(x).split()))
coindesk_df['Cleaned title'] = coindesk_df['Cleaned title'].str.replace('[^\w\s]','')
coindesk_df['Cleaned title'] = coindesk_df['Cleaned title'].apply(lambda x: " ".join(x for x in x.split() if x not in stop))

coindesk_df['Cleaned content'] = coindesk_df['Content'].apply(lambda x: " ".join(x.lower() for x in str(x).split()))
coindesk_df['Cleaned content'] = coindesk_df['Cleaned content'].str.replace('[^\w\s]','')
coindesk_df['Cleaned content'] = coindesk_df['Cleaned content'].apply(lambda x: " ".join(x for x in x.split() if x not in stop))

def space(comment):
    doc = nlp(comment)
    return " ".join([token.lemma_ for token in doc])

coindesk_df['Cleaned title'] = coindesk_df['Cleaned title'].apply(space)
coindesk_df['Cleaned content'] = coindesk_df['Cleaned content'].apply(space)
coindesk_df.head()

# Initialize FinBert
finbert = BertForSequenceClassification.from_pretrained('yiyanghkust/finbert-tone',num_labels=3)
tokenizer = BertTokenizer.from_pretrained('yiyanghkust/finbert-tone')

labels = {0:'neutral', 1:'positive',2:'negative'}

# Conduct FinBert sentiment analysis on news headlines
title_sent_val = list()
for headline in coindesk_df['Cleaned title'].to_list():
    try:
        inputs = tokenizer(headline, return_tensors="pt", padding=True)
        outputs = finbert(**inputs)[0]

        val = labels[np.argmax(outputs.detach().numpy())]   
        title_sent_val.append(val)
    except:
        title_sent_val.append(None)
        
coindesk_df['Title sentiment'] = title_sent_val

# Conduct FinBert sentiment analysis on news content
content_sent_val = list()
for content in coindesk_df['Cleaned content'].to_list():
    try:
        inputs = tokenizer(content[:512], return_tensors="pt", padding=True)
        outputs = finbert(**inputs)[0]

        val = labels[np.argmax(outputs.detach().numpy())]   
        content_sent_val.append(val)
    except:
        content_sent_val.append(None)
           
coindesk_df['Content sentiment'] = content_sent_val

# One-hot encoding of FinBert sentiment labels on news sentiment
coindesk_df['Pos title count'] = list(map(lambda x: 1 if x else 0, coindesk_df['Title sentiment']=="positive"))
coindesk_df['Neu title count'] = list(map(lambda x: 1 if x else 0, coindesk_df['Title sentiment']=="neutral"))
coindesk_df['Neg title count'] = list(map(lambda x: 1 if x else 0, coindesk_df['Title sentiment']=="negative"))

coindesk_df['Pos content count'] = list(map(lambda x: 1 if x else 0, coindesk_df['Content sentiment']=="positive"))
coindesk_df['Neu content count'] = list(map(lambda x: 1 if x else 0, coindesk_df['Content sentiment']=="neutral"))
coindesk_df['Neg content count'] = list(map(lambda x: 1 if x else 0, coindesk_df['Content sentiment']=="negative"))

coindesk_df.head()

# Aggregate tweets daily
coindesk_summary_df = coindesk_df[~(coindesk_df['Content sentiment'].isna())].groupby("Date").sum()

total_title_sent = coindesk_summary_df["Pos title count"] + coindesk_summary_df["Neu title count"] + coindesk_summary_df["Neg title count"]
coindesk_summary_df["Title proportion pos"] = coindesk_summary_df["Pos title count"] / total_title_sent
coindesk_summary_df["Title proportion neu"] = coindesk_summary_df["Neu title count"] / total_title_sent
coindesk_summary_df["Title proportion neg"] = coindesk_summary_df["Neg title count"] / total_title_sent

total_content_sent = coindesk_summary_df["Pos content count"] + coindesk_summary_df["Neu content count"] + coindesk_summary_df["Neg content count"]
coindesk_summary_df["Content proportion pos"] = coindesk_summary_df["Pos content count"] / total_content_sent
coindesk_summary_df["Content proportion neu"] = coindesk_summary_df["Neu content count"] / total_content_sent
coindesk_summary_df["Content proportion neg"] = coindesk_summary_df["Neg content count"] / total_content_sent

coindesk_summary_df.head()

coindesk_summary_df.to_csv("CoinDesk Sentiment.csv")
